feat(DTFS-6780): add typescript compatibility to dtfs-central-api
