+++ dtfs-central-api/package.json
@@ -16,15 +16,15 @@
   },
   "license": "MIT",
   "author": "UK Export Finance",
-  "main": "src/index.js",
+  "main": "src/index.ts",
   "scripts": {
     "api-test": "jest --coverage --verbose --runInBand --detectOpenHandles --config=api-test.jest.config.js ",
     "api-test-dev": "jest --verbose --runInBand --detectOpenHandles --config=api-test.jest.config.js ",
     "api-test-file": "jest --config=api-test-file.jest.config.js --runInBand --verbose --testMatch ",
     "lint": "eslint ./src",
     "lint:fix": "eslint ./src --fix",
-    "start": "node src/index.js",
-    "start:dev": "npx nodemon src/index.js",
+    "start": "npx ts-node src/index.ts",
+    "start:dev": "npx nodemon --exec npm run start",
     "unit-test": "jest --coverage --verbose --config=unit.jest.config.js",
     "unit-test-quick": "jest --config=unit.jest.config.js --testTimeout=10000"
   },
@@ -48,7 +48,8 @@
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.29.0",
     "jest": "29.5.0",
-    "supertest": "6.3.3"
+    "supertest": "6.3.3",
+    "typescript": "^5.2.2"
   },
   "engines": {
     "node": ">=18",
