--- dtfs-central-api/src/azure-env/index.js
@@ -1,8 +1,0 @@
-// Fix Azure environment variables
-Object.keys(process.env).forEach((key) => {
-  if (key.startsWith('CUSTOMCONNSTR_')) {
-    const fixedKey = key.substr('CUSTOMCONNSTR_'.length);
-    process.env[fixedKey] = process.env[key];
-    console.info('Fixed %s to %s', key, fixedKey);
-  }
-});
