+++ dtfs-central-api/src/azure-env/index.ts
@@ -1,0 +1,11 @@
+const fixAzureEnvironmentVariables = () => {
+  Object.keys(process.env).forEach((key) => {
+    if (key.startsWith('CUSTOMCONNSTR_')) {
+      const fixedKey = key.substr('CUSTOMCONNSTR_'.length);
+      process.env[fixedKey] = process.env[key];
+      console.info('Fixed %s to %s', key, fixedKey);
+    }
+  });
+};
+
+export default fixAzureEnvironmentVariables;
